refactor : rename ORM annotations and move @Mocka to mocka-core
